+++ src/pages/Loans/LoansOverview/components/LoanForm/index.tsx
@@ -1,0 +1,2 @@
+export type { LoanFormProps } from './LoanForm';
+export { LoanForm } from './LoanForm';
