+++ src/pages/Loans/LoansOverview/components/LoanModal/LoanModal.style.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Flex, Tabs, theme } from '@/component-library';
+import { Tabs, theme } from '@/component-library';
 
 const StyledTabs = styled(Tabs)`
   margin-top: ${theme.spacing.spacing6};
@@ -10,14 +10,4 @@ const StyledWrapper = styled.div`
   margin-top: ${theme.spacing.spacing6};
 `;
 
-type StyledFormWrapperProps = {
-  $showBorrowLimit: boolean;
-};
-
-const StyledFormWrapper = styled(Flex)<StyledFormWrapperProps>`
-  margin-top: ${theme.spacing.spacing8};
-  // MEMO: keep the same height across each tab
-  min-height: ${({ $showBorrowLimit }) => ($showBorrowLimit ? '32rem' : '20rem')};
-`;
-
-export { StyledFormWrapper, StyledTabs, StyledWrapper };
+export { StyledTabs, StyledWrapper };
