+++ src/utils/hooks/api/loans/use-get-account-loans-overview.tsx
@@ -26,7 +26,6 @@ interface AccountLoansOverviewData {
 interface AccountLoansOverview {
   data: AccountLoansOverviewData;
   refetch: () => void;
-
   getMaxBorrowableAmount: (
     currency: CurrencyExt,
     availableCapacity: MonetaryAmount<CurrencyExt>
@@ -51,6 +50,7 @@ const useGetAccountLoansOverview = (): AccountLoansOverview => {
   let collateralAssetsUSDValue: Big | undefined = undefined;
   let totalAccruedUSDValue: Big | undefined = undefined;
   let netYieldUSDValue: Big | undefined = undefined;
+
   let earnedRewards: MonetaryAmount<CurrencyExt> | undefined = undefined;
 
   if (lendPositions !== undefined && prices !== undefined && assets !== undefined) {
